Update intervention/image to 3.6
