+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "macocci7/php-photo-gps",
-    "version": "1.7.0",
+    "version": "1.8.0",
     "description": "Retrieves EXIF GPS data from a photo.",
     "type": "library",
     "license": "MIT",
@@ -17,7 +17,7 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "intervention/image": "^3.5",
+        "intervention/image": "^3.6",
         "nette/neon": "^3.4"
     },
     "require-dev": {
@@ -25,6 +25,6 @@
         "phpstan/phpstan": "^1.10",
         "phpmd/phpmd": "^2.15",
         "phpunit/phpunit": "^10.5",
-        "php-parallel-lint/php-parallel-lint": "1.3"
+        "php-parallel-lint/php-parallel-lint": "1.4"
     }
 }
