+++ src/Helpers/Exif.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Macocci7\PhpPhotoGps\Helper;
+namespace Macocci7\PhpPhotoGps\Helpers;
 
 use Intervention\Image\ImageManager as Image;
-use Macocci7\PhpPhotoGps\Helper\Uri;
+use Macocci7\PhpPhotoGps\Helpers\Uri;
 
 /**
  * Class for Exif Data Handling
