+++ tests/Helpers/GpsTest.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Macocci7\PhpPhotoGps\Helper;
+namespace Macocci7\PhpPhotoGps\Helpers;
 
+use Macocci7\PhpPhotoGps\Helpers\Config;
+use Macocci7\PhpPhotoGps\Helpers\Exif;
+use Macocci7\PhpPhotoGps\Helpers\Gps;
+use Nette\Neon\Neon;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
-use Macocci7\PhpPhotoGps\Helper\Gps;
-use Macocci7\PhpPhotoGps\Helper\Exif;
-use Macocci7\PhpPhotoGps\Helper\Config;
-use Nette\Neon\Neon;
 
 /**
  * @SuppressWarnings(PHPMD.CamelCaseMethodName)
